api_requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
shap
lime
gradio
python-dotenv
joblib
flask
flask-cors
pytesseract
pillow
opencv-python
tensorflow
transformers
torch
torchvision
sentence-transformers
easyocr
neo4j
networkx
tf-keras